Implement arrangement editing
This includes drag'n'drop from the pattern dock;
select, drag to copy, create variation, and delete
actions in the map; insert and delete actions for
bars; copy-paste actions for bars.

This also includes the "scroll as you drag" feature
for the bar track, and a global shortcut to pause-
unpause the playback.