Implement MIDI imports
This implementation comes with some
changes compared to the original version.
Pattern generation is probably entirely
different due to implementation detail,
drumkits are mapped to the MIDI Drumkit
instrument for better compatibility,
some convertion differences are ought to
be here and there.

Overall I think it does a similarly decent job,
but there is a lot of room for improvements
as well. All will come with more testing and
feeback, I guess.