Update the project to work around 4.3-beta1 bugs
There is an issue with Button's sizing introduced
by beta 1. It's a blocking issue, since it makes
the app's UI unusable for every icon-only button.
Thankfully, a simple workaround seems to do
the trick: just overriding all default styles for
the button makes the behavior consistent again.

This also adjusts the clear background color to
work better with the new logo. This only affects
the boot splash.

From quick but thorough testing beta 1 seems
to work fine with the GDSiON extension compiled
against an older version of godot-cpp. Eventually
the extension will also be updated, so this is a
minor concern.