Pause playback when opening file dialogs
This may help with some misbehavior from the
synth, as the native file dialog tends to lock the
main application window and stop its processing.
But the synth may try to continue and can get
caught in the middle of something.

The synth is never completely idle, unless we stop
it completely. But this may have undesirable
effects, so I'm settling on this solution for now.