Add note preview when clicking on the gutter
Also fixes a small mistake from the previous
commit which disabled previews when the
playback is paused.

This commit also finally adds a compressor to
prevent ear damaging shrieks when playing
with the LP filter. This may lead to some original
Bosca tracks to sound slightly different, but
better to sacrifice that rather than your hearing.