Extract window management into a separate global
It's a lot of code and it's arguably usable for more
than this one project. So putting it into a separate
class makes all the sense. This is already used in
Glasan FX too.

Also removes 3.0 beta compatibility code and
unusable scale options. We'll address app's scaling
properly in a follow-up commit.