Rescale the application to a more appropriate native size
This change is necessary so that the app renders at
100% scale on standard DPI screens, which removes
virtually all artifacts related to the downscaling
approach introduced earlier.

This does mean that the 100% GUI scale setting now
renders at a smaller size, so users need to adjust their
settings. But this is an overall improvements to the
UI scale.

Most of this included changes to font sizes and
spacing values, as well as some minimum size
adjustments. Some extra code changes were also
necessary in a few cases where properties used to be
needlessly hardcoded.

This commit also fixes a bug with the changes-on-exit
popup misbehaving when you press the windows
close button repeatedly.

PS. I haven't tested the results on a real device yet,
but that's the same approach that I've used in Glasan
FX, dealing with the same exact issue. So I'm already
confident in the improvements. It's even simpler here
as most of the UI is blocky and simplistic, harder to
break with rescaling.