Allow notes up to 128 units in length
Previously note lengths were capped by the
current pattern size. This doesn't match the
old behavior of Bosca Ceoil, where there is
no limit. Having no limit still seems like an
oversight to me, but I'd like to enable some
ways of compositing which were previously
possible. So for now I'm setting the upper
limit to 128, which seems reasonable.

There is already a usability concerned
introduced by this, as using the mouse wheel
to change it back and forth drastically is
hardly nice. There are keyboard shortcuts,
though, which help a bit.

Those were not documented, incidentally, so
this commit also fixes that.