[WIP] Fixup and rework the XM exporter
This ensures that the XM exporter doesn't
misbehave when dealing with the new note
size limit.

Coincidentally, there was also a related
problem with how the samples were
generated. I now adopted the same approach
to generating samples as the original Bosca
used, and which I myself used in Glasan FX
(without realizing). It's very reliable for this
specific purpose, so it's a great fit.