Add an I/O config dialog, allow to set preferred pattern size before MIDI import
This should help with note grouping when
importing compositions from MIDI. As MIDI
doesn't have a concept of patterns and is just
a stream of commands, we can't always be
sure what pattern size is appropriate. We can
approximate based on the time signature, but
that's not ideal (and besides, Bosca's own
songs may break this convention).

So a user configuration dialog is a good
compromise.