Remove the forced compression
The idea behind having it is good: prevent
ear damaging sounds from the synthesizer.
However, it's hard to configure a one-size-
fits-all compressor. Different settings suit
different instruments/voices, and trying to
reduce the peaks of a square wave we will
be destroying some other parts.

So for now I'm going commando here. The
original Bosca didn't have a compressor
either, so perhaps it's not such a big deal
here. For specific instruments we can figure
it out on the case by case basis, I guess.