Update the pattern dock more reliably
Previously we would only react on changes in the
current pattern or propagated to the Controller. Which
is not sufficient for updating the dock, especially when
undo/redo happens.

This commit adds a more standardized way of tracking
pattern list changes in the song, so the pattern dock
can be updated more appropriately.