Rework mouse wheel actions (and some others)
This deviates from the key bindings available in
the original Bosca, but for a good reason. Original
mapping uses the unmodified scroll action to
do a number of specialized things where the
context suggests normal scrolling to be expected.

This confuses some users and is generally
annoying, as you cannot easily scroll the pattern
editor or the arrangement editor. We could keep
old mapping, and add normal scrolling with a
modifier key, but that doesn't feel like a good
idea to me.

Scrolling should not be a secondary action for its
natural key/mouse combination.

So, with all that explained, here's what we're
changing:

- Unmodified mouse wheel scroll in the pattern
editor now scrolls. Same goes for the
arrangement editor.
- In the pattern editor hold Ctrl with the wheel
to change the cursor size.
- In the arrangement editor hold Shift with the
wheel to change the scale.
- In the pattern editor you can no longer erase
notes with Ctrl + LMB, as it conflicts with the
aforementioned change quite easily.
- In the arrangement editor scale can also be
changed with Ctrl + (Num) Plus/Minus. This is not
documented yet, though.

These changes address and close #15.

Not directly related, but also about shortcuts:

- Spacebar now stops instead of pausing, as it
did in the original Bosca. Closes #43.
- Ctrl + Spacebar can be used to pause like
before.
- Escape is no longer bound to quit the app,
nothing is bound to this action (but it still exists).
While we have guardrails to prevent unsaved
changes loss, you can still get into annoying
situations where you press Esc intuitively, e.g. to
close a popup, and it quits all together instead.

Additionally, the help section has been updated
to reflect these changes. But as it was getting
crowded (for a long time), I reworked it into more
sections. Arrangement help is now split into the
part about the grid, and the part about the
timeline. With more space, more details are now
explained on those pages as well.