Add support for saving and exporting songs on web
This is mostly as straightforward as it gets. Godot
exposes the `JavaScriptBridge` singleton, which
has a dedicated (hacky) method for initiating a
file download. This is exactly what I would've had
to do, so it's nice that it's there.

There are some potential engine bugs related to
FileAccess, but they seem to be easy to work
around.

Loading and importing also require some hacks
on the browser end, but this time we don't have
a dedicated engine method to rely on. Will be
implemented in the follow up.