Add support for loading and importing songs on web
This completes the file I/O part of the web
version. It still needs to be tested with browsers
and on a real server, but in principle this works
as well as it can given the limitations.

I will probably have to create a custom HTML
shell to handle browser support exceptions and
such.