Add checks and guards related to memory allocation in web exports
This is mainly relevant for Safari which has strict
budget for each page/worker. This is the main
reason why we cannot support it. Still, this may be
fixed by Apple at some point, and with these
changes we will be ready.

This also adds an early check for memory allocation
compatibility, so we can even warn the user ahead
of loading.