Improve layout of the help view
The back button has been a nuisance for quite some
time due to how height restricted these view panels
are. With this change the button is now an external
part of the panel and doesn't take space inside.

The look is borrowed from the delete area, and both
the new back button and the delete area have had
their styling adjusted.