Rework the import workflow using a new dialog
Before a similar treatment is given to the export
workflow, I'm doing a little practice on the imports.
Basically, everything is now encapsulated by an
"import master" dialog, where you both select the
file to import and adjust configuration.

While only MIDI is currently supported, so it's not
very visible, but all supported formats are simply
selectable in the file dialog. No need to choose
"Import .mid" from the button.

This new dialog also offers a great opportunity to
warn about the uncertain nature of the import in
general. And it's general presentation should make
it harder to confuse with loading/saving (there have
been precedents...)