Add online documentation to the project
This adds a system to document the project using
Markdown file, and then convert them into HTML.
The documentation build system is created using
Python with minimal dependencies.

The documentation will be built as a part of the
normal CI/CD workflow (added in a follow-up commit).
It will be hosted alongside the web version of the
app itself.

Closes #10