Fill in the file description for Windows exports
I can't find any best practices for this field
and in most examples everyone is just
using the application name. Rarely it's a
more generic description, like "XYZ driver".
But that doesn't seem particularly useful
in this case. Godot also just repeats the
application name, as a point of reference.